refactor(gui): extract Apps page builders + state into LibraryPageMixin
Move the entire Apps tab — page builder, category chip row, grid/list
populators, card/tile builders, and visibility/filter/hidden-toggle
state — off WinpodxWindow into a private mixin so main_window.py can
focus on overall window orchestration:

- LibraryPageMixin (src/winpodx/gui/_main_window_library.py)
  _build_library_page (toolbar + scroll + grid/list area),
  _build_category_chips, _set_category, _set_view, _populate_app_view,
  _populate_grid, _populate_list, _make_app_card, _make_app_tile,
  _visible_apps, _hidden_count, _refresh_hidden_button,
  _on_toggle_hidden, _filter_apps.

WinpodxWindow now reads
`class WinpodxWindow(AppCrudMixin, HeaderMixin, InfoPageMixin,
LibraryPageMixin, LogsMixin, MaintenanceMixin, PodStatusMixin, QMainWindow)`.

The hybrid Hidden-toggle UX and the search-box / category-chip /
view-toggle wiring all come along as-is. No behaviour change;
tests: 1096 passed.

main_window.py: 1371 -> 957 lines (-414). Cumulative: 2745 -> 957 (-65%).
